requirements.txt
# Core ML Libraries - Install PyTorch via Conda as per README instructions
# torch
# torchvision
# torchaudio

# Transformer Libraries
simpletransformers==0.70.0
transformers==4.42.3
sentencepiece
huggingface-hub
datasets
sacrebleu

# Fix protobuf compatibility issue (must be 3.20.x for tensorflow compatibility)
protobuf==3.20.3

# Data & Utility Libraries
pandas
tqdm

# Web App - Flask
flask
python-dotenv

# Google Gemini AI
google-generativeai

# Optional: Streamlit (legacy)
streamlit
